+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 * Adding pontos module for future releases [#60](https://github.com/greenbone/autohooks-plugin-black/pull/60)
+* Adding tests for this plugin [#89](https://github.com/greenbone/autohooks-plugin-black/pull/89)
 
 ### Changed
 Replaced pipenv with poetry for dependency management. poetry install works a bit different than pipenv install. It installs dev packages. [#51](https://github.com/greenbone/autohooks-plugin-black/pull/51)